requirements.txt
streamlit==1.37.0
openai>=1.30.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
pypdf>=4.3.1
duckduckgo-search>=6.2.11
serpapi>=0.1.5
